+++ Controllers/variantsController.js
@@ -13,9 +13,31 @@ exports.getAllVarients = async (req, res) => {
         console.log(err);
     }
 }
+
+exports.getVarientsById = async (req, res) => {
+    try {
+        const { id } = req.params;
+        const varients = await Varients.findById(id);
+        if (!varients) {
+            return res.status(404).json({ message: 'varients not found' });
+        }
+        res.status(200).json({
+            success: true,
+            data: varients,
+            message: 'Successfully varients',
+        });
+    } catch (error) {
+        console.log(error);
+        res.status(500).json({ message: error.message });
+    }
+};
+
 //Add a new varient
 exports.addVarient = async (req, res) => {
     try {
+        if (req.user.role === "user") {
+            return res.status(401).json({ message: "manna sha8ltak" });
+        }
         const {
             color,
             size,
@@ -44,6 +66,9 @@ exports.addVarient = async (req, res) => {
 // Edit Varient
 exports.editOneVarient = async (req, res) => {
     try {
+        if (req.user.role === "user") {
+            return res.status(401).json({ message: "manna sha8ltak" });
+        }
         let { id } = req.params;
         let body = req.body;
         const updateVarient = await Varients.updateOne(
@@ -64,6 +89,9 @@ exports.editOneVarient = async (req, res) => {
 //Delete One Varient
 exports.deleteVarient = async (req, res) => {
     try {
+        if (req.user.role === "user") {
+            return res.status(401).json({ message: "manna sha8ltak" });
+        }
         let { id } = req.params;
         const deleteOneVarient= await Varients.findByIdAndDelete(
             { _id: id }
